+++ src/main/java/ajou/mse/dimensionguard/domain/Room.java
@@ -37,11 +37,11 @@ public class Room extends BaseEntity {
         return of(null, RoomStatus.READY, null, null, null, null, null, null);
     }
 
-    public static Room of(Long id, RoomStatus status, LocalDateTime gameStartedAt, List<Player> players, LocalDateTime createdAt, LocalDateTime updatedAt, Integer createdBy, Integer updatedBy) {
+    public static Room of(Long id, RoomStatus status, LocalDateTime gameStartedAt, List<Player> players, LocalDateTime createdAt, LocalDateTime updatedAt, Long createdBy, Long updatedBy) {
         return new Room(id, status, gameStartedAt, players, createdAt, updatedAt, createdBy, updatedBy);
     }
 
-    private Room(Long id, RoomStatus status, LocalDateTime gameStartedAt, List<Player> players, LocalDateTime createdAt, LocalDateTime updatedAt, Integer createdBy, Integer updatedBy) {
+    private Room(Long id, RoomStatus status, LocalDateTime gameStartedAt, List<Player> players, LocalDateTime createdAt, LocalDateTime updatedAt, Long createdBy, Long updatedBy) {
         super(createdAt, updatedAt, createdBy, updatedBy);
         this.id = id;
         this.status = status;
