+++ src/main/java/ajou/mse/dimensionguard/exception/player/PlayerNotFoundByMemberAndRoomException.java
@@ -4,7 +4,7 @@ import ajou.mse.dimensionguard.exception.common.NotFoundException;
 
 public class PlayerNotFoundByMemberAndRoomException extends NotFoundException {
     
-    public PlayerNotFoundByMemberAndRoomException(Integer memberId, Integer roomId) {
+    public PlayerNotFoundByMemberAndRoomException(Long memberId, Long roomId) {
         super("memberId=" + memberId + ", roomId=" + roomId);
     }
 }
