+++ src/main/java/ajou/mse/dimensionguard/exception/room/RoomNotFoundByMemberIdException.java
@@ -3,7 +3,7 @@ package ajou.mse.dimensionguard.exception.room;
 import ajou.mse.dimensionguard.exception.common.NotFoundException;
 
 public class RoomNotFoundByMemberIdException extends NotFoundException {
-    public RoomNotFoundByMemberIdException(Integer memberId) {
+    public RoomNotFoundByMemberIdException(Long memberId) {
         super("memberId=" + memberId);
     }
 }
